requirements.txt
requests>=2.31.0
click>=8.1.0
colorama>=0.4.6
pyyaml>=6.0
tqdm>=4.66.0
python-dateutil>=2.8.2